server/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
flask==2.0.1
flask-cors==3.0.10
pymongo==4.3.3
scikit-learn==1.2.2
torch==2.0.0
elasticsearch==7.17.9
psycopg2-binary==2.9.6
python-dotenv==0.19.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-asyncio==0.21.0
aiohttp==3.8.4
python-jose==3.3.0
bcrypt==4.0.1